archadium/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "archadium"
version = "0.1.0"
description = "A text adventure game with rich terminal effects"
requires-python = ">=3.11"
dependencies = [
    "rich>=13.0",
    "pyyaml>=6.0",
]

[tool.setuptools.packages.find]
include = ["archadium*"]

[tool.setuptools.package-data]
archadium = ["data/**/*.yaml", "data/**/*.txt"]